requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
langchain>=0.2.0
langgraph>=0.1.0
groq>=0.9.0
structlog>=24.1.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
websockets>=12.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
chromadb>=1.0.0
